mv scripts out of package dir; fix regression: import in JTAGCom
Scripts should not resided in a package directory. If they are in
the source folder of a package installed on your system, they prefer
local versions of the modules over the installed ones often resulting
in confusing behaviour.

The relative import removed from JTAGCom in f33cb4a was done because
of this erratic behaviour. We go back to the relative import in this
commit.